Plan 9 from Bell Labs 2003-12-20

+++ sys/src/ape/lib/ap/plan9/_buf.c
@@ -384,7 +384,8 @@ alarmed(int v)
 static void
 _killtimerproc(void)
 {
-	kill(timerpid, SIGKILL);
+	if(timerpid > 0)
+		kill(timerpid, SIGKILL);
 }
 
 static void
